+++ test/conformance/testdata/conformance.yaml
@@ -420,6 +420,20 @@
     send events when scheduling and running a Pod.
   release: v1.9
   file: test/e2e/node/events.go
+- testname: Pod Eviction, Toleration limits
+  codename: '[k8s.io] [sig-node] NoExecuteTaintManager Multiple Pods [Serial] evicts
+    pods with minTolerationSeconds [Disruptive] [Conformance]'
+  description: In a multi-pods scenario with tolerationSeconds, the pods MUST be evicted
+    as per the toleration time limit.
+  release: v1.16
+  file: test/e2e/node/taints.go
+- testname: Taint, Pod Eviction on taint removal
+  codename: '[k8s.io] [sig-node] NoExecuteTaintManager Single Pod [Serial] removing
+    taint cancels eviction [Disruptive] [Conformance]'
+  description: The Pod with toleration timeout scheduled on a tainted Node MUST not
+    be evicted if the taint is removed before toleration time ends.
+  release: v1.16
+  file: test/e2e/node/taints.go
 - testname: Pods, QOS
   codename: '[k8s.io] [sig-node] Pods Extended [k8s.io] Pods Set QOS Class should
     be set on Pods with matching resource requests and limits for memory and cpu [Conformance]'
@@ -1694,20 +1708,6 @@
     validate the pod resources are applied to the Limitrange
   release: v1.18
   file: test/e2e/scheduling/limit_range.go
-- testname: Pod Eviction, Toleration limits
-  codename: '[sig-scheduling] NoExecuteTaintManager Multiple Pods [Serial] evicts
-    pods with minTolerationSeconds [Disruptive] [Conformance]'
-  description: In a multi-pods scenario with tolerationSeconds, the pods MUST be evicted
-    as per the toleration time limit.
-  release: v1.16
-  file: test/e2e/scheduling/taints.go
-- testname: Taint, Pod Eviction on taint removal
-  codename: '[sig-scheduling] NoExecuteTaintManager Single Pod [Serial] removing taint
-    cancels eviction [Disruptive] [Conformance]'
-  description: The Pod with toleration timeout scheduled on a tainted Node MUST not
-    be evicted if the taint is removed before toleration time ends.
-  release: v1.16
-  file: test/e2e/scheduling/taints.go
 - testname: Scheduler, resource limits
   codename: '[sig-scheduling] SchedulerPredicates [Serial] validates resource limits
     of pods that are allowed to run  [Conformance]'
