+++ test/e2e/node/taints.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package scheduling
+package node
 
 import (
 	"context"
