+++ test/integration/BUILD
@@ -61,6 +61,7 @@ filegroup(
         "//test/integration/master:all-srcs",
         "//test/integration/metrics:all-srcs",
         "//test/integration/namespace:all-srcs",
+        "//test/integration/node:all-srcs",
         "//test/integration/objectmeta:all-srcs",
         "//test/integration/openshift:all-srcs",
         "//test/integration/pods:all-srcs",
