+++ test/integration/node/OWNERS
@@ -1,0 +1,6 @@
+# See the OWNERS docs at https://go.k8s.io/owners
+
+approvers:
+- sig-node-approvers
+reviewers:
+- sig-node-reviewers
